changes to report stream callback to add queue name
