+++ backend/src/main/java/gov/cdc/usds/simplereport/api/model/ReportStreamCallbackRequest.java
@@ -8,4 +8,5 @@ public class ReportStreamCallbackRequest {
   private UUID testEventInternalId;
   private Boolean isError;
   private String details;
+  private String queueName;
 }
