+++ ops/services/app_functions/report_stream_batched_publisher/functions/QueueBatchedReportStreamUploader/index.ts
@@ -10,7 +10,7 @@ import {
   reportExceptions,
   deleteSuccessfullyParsedMessages,
 } from "../common/queueHandlers";
-import { ReportStreamResponse } from "../common/rs-response";
+ import { ReportStreamResponse } from "../common/types";
 
 const {
   REPORT_STREAM_URL,
