+++ ops/services/app_functions/report_stream_batched_publisher/functions/ReportStreamExceptionHandler/index.ts
@@ -1,13 +1,13 @@
 import { AzureFunction, Context } from "@azure/functions";
 import fetch, { Headers } from "node-fetch";
 import { ENV } from "../config";
-import { SimpleReportReportStreamResponse } from "../common/rs-response";
+import { ReportStreamCallbackRequest } from "../common/types";
 
 const { SIMPLE_REPORT_CB_TOKEN, SIMPLE_REPORT_CB_URL } = ENV;
 
 const queueTrigger: AzureFunction = async function (
   context: Context,
-  message: SimpleReportReportStreamResponse
+  message: ReportStreamCallbackRequest // This type needs to be changed because the objects places in here will be a mutation of the SimpleReportReportStreamResponse + queue name
 ): Promise<void> {
   const headers = new Headers({
     Accept: "application/json",
@@ -23,7 +23,7 @@ const queueTrigger: AzureFunction = async function (
     throw new Error(`${result.status}: ${await result.text()}`);
   }
   context.log(
-    `Successfully processed exception for record ${message.testEventInternalId}`
+    `Successfully processed exception for record ${message.testEventInternalId} from queue ${message.queueName}`
   );
 };
