+++ ops/services/app_functions/report_stream_batched_publisher/functions/common/reportingHandlers.test.ts
@@ -13,7 +13,7 @@ import {
 } from "./reportingHandlers";
 import { SimpleReportTestEvent } from "../FHIRTestEventReporter/dataHandlers";
 // import { uploaderVersion } from "../config";
-import { ReportStreamResponse } from "./rs-response";
+import { ReportStreamResponse } from "./types";
 
 jest.mock(
   "node-fetch",
